button refactor
