+++ .changeset/tricky-news-notice.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/button': patch
+---
+
+Refactors Button component internals
