+++ packages/button/src/ButtonIcon/ButtonIcon.tsx
@@ -4,8 +4,8 @@ import { css, cx } from '@leafygreen-ui/emotion';
 import { getTheme, Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 
-import { ButtonClassName } from './styles';
-import { ButtonProps, Size, Variant } from './types';
+import { ButtonClassName } from '../styles';
+import { ButtonProps, Size, Variant } from '../types';
 
 const baseIconStyle: Record<Theme, Record<Variant, string>> = {
   [Theme.Light]: {
