+++ packages/button/src/index.ts
@@ -1,3 +1,3 @@
-export { Button as default } from './Button';
+export { Button as default } from './Button/Button';
 export type { ButtonProps } from './types';
 export { FontSize, Size, Variant } from './types';
